requirements.txt
streamlit
pdfplumber
pdf2image
Pillow
PyMuPDF
moviepy
opencv-python-headless
requests
imageio-ffmpeg